+++ src/main/java/com/example/waplan/login/api/LoginApi.java
@@ -4,9 +4,8 @@ import com.example.waplan.login.application.LoginService;
 import com.example.waplan.login.application.dto.FindUserIdRequest;
 import com.example.waplan.login.application.dto.ResetPasswordRequest;
 import com.example.waplan.login.application.dto.SignUpRequest;
-import com.example.waplan.mail.application.MailService;
-import com.example.waplan.mail.domain.EmailMessage;
-import com.example.waplan.user.application.UserService;
+import com.example.waplan.login.application.MailService;
+import com.example.waplan.login.domain.EmailMessage;
 import jakarta.validation.Valid;
 import java.net.URI;
 import lombok.RequiredArgsConstructor;
