+++ src/main/java/com/example/waplan/login/application/MailService.java
@@ -1,12 +1,10 @@
-package com.example.waplan.mail.application;
+package com.example.waplan.login.application;
 
-import com.example.waplan.login.application.LoginService;
-import com.example.waplan.mail.domain.EmailMessage;
+import com.example.waplan.login.domain.EmailMessage;
 import jakarta.mail.MessagingException;
 import jakarta.mail.internet.MimeMessage;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.mail.javamail.JavaMailSender;
 import org.springframework.mail.javamail.MimeMessageHelper;
 import org.springframework.stereotype.Service;
